Moved dsp functions, as well as Knob widget, to a separate repo for reusability
