+++ Cargo.toml
@@ -19,15 +19,12 @@ image = {version = "0.25.5", features = ["png"]}
 lazy_static = "1.4.0"
 serde = { version = "1.0", features = ["derive"] }
 
-dsp_core = {path = "dsp_core"}
+plugin-utils = { git = "https://github.com/Garneek/plugin-utils.git" }
 
 [workspace]
 members = [
   "xtask",
-  "dsp_core",
-  
 ]
-
 
 [lib]
 path = "src/lib.rs"
