--- dsp_core/Cargo.toml
@@ -1,10 +1,0 @@
-[package]
-name = "dsp_core"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-nih_plug = { git = "https://github.com/ardura/nih-plug.git", rev = "de315f902d7b5a75f80bc9d4164e4485bf2b34a1", features = ["assert_process_allocs"] }
-
-[lib]
-path = "src/lib.rs"
