+++ src/editor/style.rs
@@ -1,5 +1,5 @@
-use crate::editor::KnobLayout;
-use crate::editor::KnobPreset;
+use plugin_utils::egui_utils::KnobLayout;
+use plugin_utils::egui_utils::KnobPreset;
 
 use nih_plug_egui::egui::Color32;
 
@@ -7,16 +7,28 @@ pub(crate) const WIDTH: u32 = 400;
 pub(crate) const HEIGHT: u32 = 488;
 
 // Ferra color palette
+
+#[allow(dead_code)]
 pub(crate) const FERRA_NIGHT: Color32 = Color32::from_rgb(42, 41, 45);
+#[allow(dead_code)]
 pub(crate) const FERRA_ASH: Color32 = Color32::from_rgb(55, 53, 57);
+#[allow(dead_code)]
 pub(crate) const FERRA_UMBER: Color32 = Color32::from_rgb(77, 66, 75);
+#[allow(dead_code)]
 pub(crate) const FERRA_BARK: Color32 = Color32::from_rgb(111, 93, 99);
+#[allow(dead_code)]
 pub(crate) const FERRA_MIST: Color32 = Color32::from_rgb(209, 209, 224);
+#[allow(dead_code)]
 pub(crate) const FERRA_SAGE: Color32 = Color32::from_rgb(177, 182, 149);
+#[allow(dead_code)]
 pub(crate) const FERRA_BLUSH: Color32 = Color32::from_rgb(254, 205, 178);
+#[allow(dead_code)]
 pub(crate) const FERRA_CORAL: Color32 = Color32::from_rgb(255, 160, 122);
+#[allow(dead_code)]
 pub(crate) const FERRA_ROSE: Color32 = Color32::from_rgb(246, 182, 201);
+#[allow(dead_code)]
 pub(crate) const FERRA_EMBER: Color32 = Color32::from_rgb(224, 107, 117);
+#[allow(dead_code)]
 pub(crate) const FERRA_HONEY: Color32 = Color32::from_rgb(245, 215, 110);
 
 pub(crate) const KNOB_WIDTH: f32 = 38_f32 * 2_f32 + 4.75_f32 + 16_f32;
@@ -43,7 +55,7 @@ pub(crate) const KNOB_PRESET: KnobPreset = KnobPreset {
     readable_box: None,
     background_radius: Some(BACKGROUND_ROUNDING),
     background_opacity: Some(BACKGROUND_OPACITY),
-    layout: KnobLayout::Vertical,
+    layout: Some(KnobLayout::Vertical),
     arc_start: None,
     arc_end: None,
 };
