--- src/editor/widgets.rs
@@ -1,8 +1,0 @@
-use nih_plug_egui::egui;
-
-pub use nih_plug_egui::egui::CentralPanel;
-
-mod knob;
-pub use knob::ArcKnob;
-pub use knob::KnobLayout;
-pub use knob::KnobPreset;
